Code refactoring
The code is grouped from different files so as not to overload the App.js file

A file was created for the styles, a json file to contain the data and a file that contains the Item component.
